physics_test confirms there is a problem with angular velocity resolution with continuous collision: it completely stops all motion
