+++ app/app.c
@@ -1,9 +1,29 @@
 #include "app.h"
+#include "camera.h"
+#include "../systems/default_pass.h"
+#include "../systems/gravity.h"
+
+// wrappers
+#define BOX_MESH 0
+#define ICOSAHEDRON_MESH BOX_MESH + 1
+#define ICOSPHERE_MESH ICOSAHEDRON_MESH + 1
+
+typedef struct application_data_t
+{
+    view_camera camera;
+} application_data;
+
+static application_data app_data = { 0 };
 
 //#define STRESS_TEST
+#ifndef STRESS_TEST
+#define PHYSICS_TEST
+#endif
 
 #ifdef STRESS_TEST
 #include "stress_test.h"
+#elif defined(PHYSICS_TEST)
+#include "physics_test.h"
 #else
 #include "space_sim.h"
 #endif
@@ -15,9 +35,46 @@ if(app_return != SUCCESS) return app_return;\
 
 return_code app_run()
 {
+    // init camera
+    float d = 6.0f;
+    init_camera(dm_vec3_set(d,d,d), dm_vec3_set(-1,-1,-1), 0.1f, 100.0f, 45.0f, 0.3f, 5.0f, DM_SCREEN_WIDTH, DM_SCREEN_HEIGHT, &app_data.camera);
+    
+    // link camera view_proj to debug draw pass
+    dm_debug_render_set_view_proj(&app_data.camera.view_proj);
+    dm_debug_render_set_inv_view(&app_data.camera.inv_view);
+    dm_debug_render_set_far_plane(&app_data.camera.far_plane);
+    
+    // mesh data
+    float* positions = NULL;
+    float* normals = NULL;
+    float* tex_coords = NULL;
+    uint32_t* indices = NULL;
+    uint32_t num_vertices=0, num_indices=0;
+    
+    dm_render_handle meshes[3];
+    uint32_t num_meshes=0;
+    // box mesh
+    dm_geometry_box(&positions, &normals, &tex_coords, &indices, num_vertices, &num_vertices, &num_indices, &meshes[num_meshes++]);
+    
+    // icosahedorn mesh
+    dm_geometry_icosahedron(&positions, &normals, &tex_coords, &indices, num_vertices, &num_vertices, &num_indices, &meshes[num_meshes++]);
+    
+    // icosphere mesh
+    dm_geometry_icosphere(4, &positions, &normals, &tex_coords, &indices, num_vertices, &num_vertices, &num_indices, &meshes[num_meshes++]);
+    
+    // submit data
+    if(!default_pass_init(positions, normals, tex_coords, num_vertices, indices, num_indices, meshes, DM_ARRAY_LEN(meshes), &app_data.camera)) return INIT_FAIL;
+    
+    dm_free(positions);
+    dm_free(normals);
+    dm_free(tex_coords);
+    dm_free(indices);
+    
     ////////////////////////////////////
 #ifdef STRESS_TEST
     APP_FUNC_CHECK(stress_test_init());
+#elif defined(PHYSICS_TEST)
+    APP_FUNC_CHECK(physics_test_init());
 #else
     APP_FUNC_CHECK(space_sim_init());
 #endif
@@ -46,6 +103,12 @@ return_code app_run()
         
         if(dm_input_key_just_pressed(DM_KEY_P)) dm_physics_toggle_pause();
         
+        // resize camera
+        uint32_t width = DM_SCREEN_WIDTH;
+        uint32_t height = DM_SCREEN_HEIGHT;
+        
+        resize_camera(width, height, &app_data.camera);
+        
         // render
         dm_timer render_timer = { 0 };
         dm_timer_start(&render_timer);
@@ -55,6 +118,8 @@ return_code app_run()
         //////////////////////////////////////
 #ifdef STRESS_TEST
         APP_FUNC_CHECK(stress_test_render());
+#elif defined(PHYSICS_TEST)
+        APP_FUNC_CHECK(physics_test_render());
 #else
         APP_FUNC_CHECK(space_sim_render());
 #endif
@@ -72,9 +137,11 @@ return_code app_run()
         
         //////////////////////////////////////
 #ifdef STRESS_TEST
-        APP_FUNC_CHECK(stress_test_update());
+        APP_FUNC_CHECK(stress_test_update(&app_data.camera));
+#elif defined(PHYSICS_TEST)
+        APP_FUNC_CHECK(physics_test_update(&app_data.camera));
 #else
-        APP_FUNC_CHECK(space_sim_update());
+        APP_FUNC_CHECK(space_sim_update(&app_data.camera));
 #endif
         //////////////////////////////////////
