+++ app/camera.h
@@ -18,5 +18,6 @@ void update_camera(float delta_time, view_camera* camera);
 void resize_camera(uint32_t width, uint32_t height, view_camera* camera);
 
 void track_camera(dm_vec3 pos, float angle, view_camera* camera);
+void set_camera_pos(dm_vec3 pos, view_camera* camera);
 
 #endif //CAMERA_H
